request_to_string() implementation in mc_api class
